update cluster states page and related pages
