+++ docs/_docs/cluster-states.adoc
@@ -1,21 +1,61 @@
 = Cluster States
 
+:javaFile: {javaCodeDir}/ClusterAPI.java
 
-You cluster can be in one of the three states.
+== Overview
+
+Your cluster can be in one of the three states: `ACTIVE`, `ACTIVE_READ_ONLY`, and `INACTIVE`.
+
+When you start a pure in-memory cluster (no persistent data regions) for the first time, the cluster is in the `ACTIVE` state.
+When you start a cluster with persistent data regions for the first time, the cluster is `INACTIVE`.
 
 
 * `INACTIVE`: All operations are prohibited.
 +
 --
-The default state of the cluster when you start it for the first time
-
-When you first start your cluster, pure in-memory cluster (no persistent data regions)
-
 When you change the cluster state from active to `INACTIVE` (deactivation), the cluster deallocates all memory resources.
 
 include::includes/note-on-deactivation.adoc[]
 
 --
-* `ACTIVE`: This is the normal mode of the cluster. You can change change the cluster topology and execute any operations.
+* `ACTIVE`: This is the normal mode of the cluster. You can execute any operation.
 
-* `ACTIVE_READ_ONLY`: The read-only mode. The operations that modify data in the cluster are prohibited.
+* `ACTIVE_READ_ONLY`: The read-only mode. Only read operations are allowed.
+
+== Changing Cluster State
+
+You can change the cluster state in multiple ways:
+
+* Control script:
++
+[source, shell]
+----
+control.sh --set-state ACTIVE_READ_ONLY
+----
+
+* REST command:
++
+--
+
+[source, url]
+----
+http://localhost:8080/ignite?cmd=setstate&state=ACTIVE_READ_ONLY
+----
+
+
+--
+* Programmatically
++
+[tabs]
+--
+tab:Java[]
+
+[source, java]
+----
+include::{javaFile}[tags=change-state, indent=0]
+----
+
+
+tab:C#/.NET[]
+tab:C++[]
+--
