+++ docs/_docs/restapi.adoc
@@ -310,16 +310,60 @@ http://host:port/ignite?cmd=version
 }
 ----
 
-=== Activate
-
-Activates the cluster.
+=== Cluster State
+Returns the current link:cluster-states[state of the cluster].
 
 *Request:*::
 +
 [source,shell]
 ----
-http://host:port/ignite?cmd=activate
+http://host:port/ignite?cmd=state
 ----
+
+*Response:*::
++
+Returns `true` if the cluster is active. Returns `false` if the cluster in inactive.
++
+[source,json]
+----
+{
+  "successStatus":0,
+  "error":null,
+  "sessionToken":null,
+  "response": "ACTIVE_READ_ONLY"
+}
+----
+
+
+=== Change Cluster State
+
+
+*Request:*::
++
+--
+[source,shell]
+----
+http://host:port/ignite?cmd=setstate&state={new_state}
+----
+
+[{request_table_props}]
+|===
+|Parameter
+|Type
+|Optional
+|Description
+|Example
+
+|`state` | String| No a| New cluster state. One of the values:
+
+* `ACTIVE`: active state,
+* `ACTIVE_READ_ONLY`: read only state,
+* `INACTIVE`: the cluster is deactivated.
+
+include::includes/note-on-deactivation.adoc[]
+|
+|===
+--
 
 *Response:*::
 +
@@ -329,11 +373,11 @@ http://host:port/ignite?cmd=activate
   "successStatus":0,
   "error":null,
   "sessionToken":null,
-  "response":"activate started"
+  "response":"setstate done"
 }
 ----
 
-
+////
 === Deactivate
 Starts the deactivation process for a persistence-enabled cluster.
 
@@ -358,31 +402,8 @@ http://host:port/ignite?cmd=deactivate
   "response":"deactivate started"
 }
 ----
+////
 
-
-=== Current State
-Returns the current state (active/inactive) of the cluster.
-
-*Request:*::
-+
-[source,shell]
-----
-http://host:port/ignite?cmd=currentstate
-----
-
-*Response:*::
-+
-Returns `true` if the cluster is active. Returns `false` if the cluster in inactive.
-+
-[source,json]
-----
-{
-  "successStatus":0,
-  "error":null,
-  "sessionToken":null,
-  "response":true
-}
-----
 
 === Increment
